feat: add bun lockfile and address Node20 TS errors
